Ditched Drift, add a starboard module, add proper migrations
Yes, this commit is big, rewriting the codebase to remove drift generates a large diff.
This also add a starboard module, because why not? And adds proper migration handling.
I also overhauled some stuff, made converters for more types and handle auto complete via converters and not the `@Autocomplete` annotation.
The api/website are currently heavy WIP so these are not going to be revised (or described here)

I plan to commits more often, I was very (and still am tbh) overwhelmed recently, but this should be doable (well, I hope so 💀)